requirements.txt
openai
langchain
wikipedia
python-dotenv
tqdm

